refactor: refactor style for mainPage
