+++ src/components/main/MainImage.ts
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+const MainImage = styled.img`
+  width: 18rem;
+  height: 14rem;
+  @media all and (max-width: 767px) {
+    width: 15rem;
+    height: 13rem;
+  }
+`;
+
+export { MainImage };
