+++ src/components/result/ResultImage.ts
